Moved flyctl to src/flyctl
